All pages can now be served using flask, etc.
All pages are modified so that they can be served. The endpoints
are:-

/ :- homepage
/user/<userid> :- user profile pages
/artist/<artistid> :- artist pages
/song/<songid> :- Song profiles

Anything else = 404 not found error.

All the code is now refactored, please take a look.

Also test.py can be changed if you know how to write a unit test. That
file is there just to mess around with Travis.

A lot of changes are visible, others not so visible.